clean direct preference algos
